Create diskspace.sh
Bash shell script to monitor or watch the disk space on your server, it will send an email to $ADMIN, if the (free available) percentage of space is >= 90% or you can change the value of it. Make sure to change or put your email on $ADMIN so that you can get email. 
Then apply it on your crontab every 10-15 minutes